Initial splitting of files
