+++ src/components/Plan.scss
@@ -35,90 +35,3 @@
   position: relative;
   top: 1px;
 }
-
-
-.date-column {
-  flex: 1;
-  text-align: right;
-  margin-right: 20px;
-}
-
-.date-string {
-  font-weight: bold;
-}
-
-.workout.edit .date-string {
-  /* Align date with total distance */
-  padding-top: 2px;
-}
-
-.description-column {
-  flex: 2;
-}
-
-.edit .description-column .my-row:not(:last-child),
-.edit .date-column .my-row {
-  margin-bottom: 10px;
-}
-
-.my-row input {
-  margin-bottom: 0;
-}
-
-.description-input {
-  width: 100%;
-  display: block;
-}
-
-.workout {
-  display: flex;
-  flex-direction: row;
-  padding: 10px 0;
-}
-
-.workout.edit:not(:last-child) {
-  border-bottom: 1px solid #d2d0ce;
-}
-
-.goal-race, .actions-container {
-  margin-bottom: 1.25em;
-  font-size: 1.17em;
-}
-
-input.total-distance-input {
-  /* 3 chars + up/down toggle in some browsers  */
-  width: 5em;
-  margin: 0 5px;
-}
-
-/* Align actions with the description input */
-.edit-workout-action-container {
-  padding-top: 5px;
-}
-
-.workout:not(:focus-within):not(:hover) .edit-workout-action-container {
-  visibility: hidden;
-}
-
-.workout-action-button:last-child {
-  padding-right: 0;
-}
-
-@media (max-width: 800px) {
-  .workout {
-    display: block;
-  }
-
-  .date-column {
-    margin-right: 0px;
-    text-align: left;
-  }
-
-  .workout:not(:focus-within):not(:hover) .edit-workout-action-container {
-    visibility: visible;
-  }
-
-  .edit-workout-action-container {
-    padding-top: 0px;
-  }
-}
